requirements.txt
# Restaurant SaaS - Python Dependencies
# Production-ready ML Engineering Stack

# ============================================
# Web Framework
# ============================================
fastapi>=0.124.2
uvicorn[standard]>=0.28.0
python-multipart>=0.0.9

# ============================================
# Database
# ============================================
sqlalchemy[asyncio]>=2.0.45
asyncpg>=0.29.0
aiosqlite>=0.20.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Redis
redis>=5.0.0

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# ============================================
# Data Processing
# ============================================
pandas>=2.3.3
numpy>=1.26.0
scipy>=1.12.0

# ============================================
# Machine Learning - Core
# ============================================
scikit-learn>=1.8.0
lightgbm>=4.6.0
xgboost>=3.1.2

# ============================================
# MLOps - Experiment Tracking & Model Registry
# ============================================
mlflow>=3.5.1
optuna>=3.6.0
optuna-integration>=3.6.0

# ============================================
# Data Validation & Quality
# ============================================
pandera>=0.18.0
great-expectations>=0.18.0

# ============================================
# Model Explainability
# ============================================
shap>=0.45.0
lime>=0.2.0

# ============================================
# Embeddings & NLP
# ============================================
sentence-transformers>=2.6.0
anthropic>=0.75.0

# Vector Database Support
pgvector>=0.2.5

# ============================================
# HTTP Client
# ============================================
httpx>=0.27.0

# ============================================
# Configuration
# ============================================
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# ============================================
# AI/LLM Providers
# ============================================
groq>=0.11.0

# ============================================
# Visualization
# ============================================
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.19.0

# ============================================
# Testing
# ============================================
pytest>=8.1.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
hypothesis>=6.99.0

# ============================================
# Notebooks
# ============================================
jupyter>=1.0.0
ipykernel>=6.29.0
nbformat>=5.10.0

# ============================================
# Development
# ============================================
black>=24.2.0
isort>=5.13.0
mypy>=1.9.0
ruff>=0.3.0
pre-commit>=3.6.0
